requirements.txt
openai-agents>=0.14.4
openai>=2.32.0
python-dotenv>=1.2.2
pandas>=2.0
fastmcp>=3.2.4
fastapi>=0.136.0
uvicorn>=0.45.0
